requirements.txt
black==23.11.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
cssutils==2.9.0
idna==3.4
lxml==4.9.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
premailer==3.10.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
yagmail==0.15.293
